Refactor resource syncer code into smaller files
